+++ packages/component-library/src/components/layout/mt-empty-state/mt-empty-state.interactive.stories.ts
@@ -1,7 +1,8 @@
-import type { Meta, StoryFn } from "@storybook/vue3";
+import type { Meta, StoryFn, StoryObj } from "@storybook/vue3";
 import { within, expect } from "@storybook/test";
 import MtEmptyState from "./mt-empty-state.vue";
 import meta, { Default, Extended } from "./mt-empty-state.stories";
+import MtButton from "@/components/form/mt-button/mt-button.vue";
 
 export default {
   ...meta,
@@ -38,4 +39,23 @@ VisualTestExtendedPage.play = async ({ canvasElement }) => {
 
   expect(canvas.getByText(Extended.args?.linkText)).toBeVisible();
   expect(canvas.getByText(Extended.args?.buttonText)).toBeVisible();
+};
+
+export const VisualTestExtendedWithSlot: StoryObj<typeof MtEmptyState> = {
+  args: {
+    ...Extended.args,
+  },
+  render: () => ({
+    components: { MtEmptyState, MtButton },
+    template: `
+      <mt-empty-state
+        icon="solid-chart-line-arrow"
+        headline="Hello, world"
+        description="This is some custom description"
+      >
+        <template #button>
+          <mt-button variant="primary">This is a custom button</mt-button>
+        </template>
+      </mt-empty-state>`,
+  }),
 };