+++ packages/admin-sdk/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [1147b22]
+  - @shopware-ag/meteor-component-library@4.12.1
+
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [215eec6]
+  - @shopware-ag/meteor-component-library@4.12.0
+
+## 6.0.0
+
+### Major Changes
+
+- 0e794ae: Add `uiMediaModalOpen` method
+  Add `media` option for parameter `entity`, `view` option for parameter `view` of actionButton method
+
+### Minor Changes
+
+- 9956281: Add textContent property to modal triggering instead of only locationId
+
+### Patch Changes
+
+- Updated dependencies [a85dee7]
+  - @shopware-ag/meteor-component-library@4.11.2
+
 ## 5.9.2
 
 ### Patch Changes